(/usr/share/games/fortunes/tao) % A good walker leaves no tracks; A good speaker makes no slips; A good reckoner needs no tally. A good door needs no lock, Yet no one can open it. Good binding requires no knots, Yet no one can loosen it.
Therefore the sage takes care of all men
And abandons no one.
He takes care of all things
And abandons nothing.

This is called "following the light."

What is a good man?
A teacher of a bad man.
What is a bad man?
A good man's charge.
If the teacher is not respected,
And the student not cared for,
Confusion will arise, however clever one is.
This is the crux of mystery.